ml-sales-prediction-system/requirements.txt
# Core Framework
Flask==3.0.3
flask-cors==5.0.0
flask-restful==0.3.10
flasgger==0.9.7.1

# WSGI Server
gunicorn==22.0.0

# Machine Learning
scikit-learn>=1.5.0
xgboost>=2.1.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.4.0

# Data Validation
marshmallow==3.21.3

# Database
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.30

# Environment
python-dotenv==1.0.1

# Monitoring
prometheus-client==0.20.0

# Security
PyJWT==2.8.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0

# Logging
colorlog==6.8.2

# Rate Limiting
Flask-Limiter>=3.7.0
